requirements/requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.7.2/constraints-3.11.txt"

apache-airflow-providers-snowflake==5.0.1
apache-airflow-providers-mysql==5.3.1
pandas==2.1.1
scikit-learn==1.2.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-api-core==2.12.0
google-api-python-client==2.102.0
apache-airflow-providers-slack==8.1.0
retry==0.9.2